# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dastkar Exhibition API",
    "version": "1.0.0",
    "description": "This is the swagger specification for Dastkar Exhibition System"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": " Local host server details"
    }
  ],
  "paths": {
...
    "/login": {
      "post": {
        "description": "Login to the application",
        "operationId": "loginUser",
        "tags": ["Session Management"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDetails" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull Login",
            "headers": {
              "session": {
                "description": "header value containing the JWT for the session",
                "schema": { "type": "string" }
              },
              "Set-Cookie": {
                "description": "set cookie with the jwt for the session",
                "schema": {
                  "type": "string",
                  "example": "sessionid=abcde12345; Path=/; HttpOnly"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status" }
              }
            }
          },
          "204": {
            "description": "Successful login",
            "headers": {
              "session": {
                "description": "header value containing the JWT for the session",
                "schema": { "type": "string" }
              },
              "Set-Cookie": {
                "description": "set cookie with the jwt for the session",
                "schema": {
                  "type": "string",
                  "example": "sessionid=abcde12345; Path=/; HttpOnly"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status" }
              }
            }
          },
          "401": {
            "description": "Username and password combination is incorrect",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "404": {
            "description": "Username not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
    "/whoami": {
      "get": {
        "description": "Get logged in user's username",
        "operationId": "whoami",
        "tags": ["Session Management"],
        "responses": {
          "200": {
            "description": "success response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/WhoAmIResponse" }
              }
            }
          },
          "404": { "description": "no user found response" }
        }
      }
    },
...
    "/user": {
      "get": {
        "description": "Fetch all the users",
        "tags": ["User Management"],
        "operationId": "fetchAllUsers",
        "x-middleware": ["roleMiddleware"],
        "x-role": ["dastkar-user-readAll", "dastkar-app-admin"],
        "responses": {
          "200": {
            "description": "successfull retrieval",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/UserDetails" }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Create user of the application",
        "tags": ["User Management"],
        "operationId": "createUser",
        "x-middleware": ["roleMiddleware"],
        "x-role": ["dastkar-user-create", "dastkar-app-admin"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserDetails" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User creation successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status" }
              }
            }
          }
        }
      }
    },
    "/user/{username}": {
      "parameters": [{ "$ref": "#/components/parameters/username" }],
      "get": {
        "description": "Fetch all the users",
        "tags": ["User Management"],
        "operationId": "fetchUser",
        "x-middleware": ["roleMiddleware"],
        "x-role": ["dastkar-user-read", "dastkar-app-admin"],
        "responses": {
          "200": {
            "description": "successfull retrieval",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserDetails" }
              }
            }
          }
        }
      },
      "post": {
        "description": "update user of the application",
        "tags": ["User Management"],
        "operationId": "updateUser",
        "x-middleware": ["roleMiddleware"],
        "x-role": ["dastkar-user-update", "dastkar-app-admin"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserDetails" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User creation successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status" }
              }
            }
          }
        }
      },
      "delete": {
        "description": "delete user",
        "tags": ["User Management"],
        "operationId": "removeUser",
        "x-middleware": ["roleMiddleware"],
        "x-role": ["dastkar-user-delete", "dastkar-app-admin"],
        "responses": {
          "200": {
            "description": "successfull removal",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Status" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
...
      "username": {
        "in": "path",
        "name": "username",
        "required": true,
        "schema": { "type": "string" },
        "description": "username of the user"
      }
    },
    "schemas": {
...
      "LoginDetails": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "example": "johndoe" },
          "password": { "type": "string", "example": "johndoe1234" }
        }
      },
      "WhoAmIResponse": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "example": "johndoe" },
          "firstname": { "type": "string", "example": "john" },
          "lastname": { "type": "string", "example": "doe" }
        }
      },
...
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "string", "example": "User-101" },
          "message": { "type": "string", "example": "User already exists" },
          "description": { "type": "string", "example": "User already exists" }
        }
      },
      "Status": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": {
            "type": "string",
            "example": "Successfuly executed command"
          }
        }
      },
      "UserDetails": {
        "type": "object",
        "required": ["identifier"],
        "properties": {
          "identifier": {
            "type": "object",
            "properties": {
              "username": { "type": "string", "example": "johndoe123" },
              "firstName": { "type": "string", "example": "john" },
              "lastName": { "type": "string", "example": "doe" },
              "email": { "type": "string", "example": "johndoe@email.com" },
              "mobileNumber": { "type": "string", "example": "+61123123123" }
            }
          },
          "password": { "type": "string", "example": "johndoe1234" },
          "oldPassword": { "type": "string", "example": "johndoe1234" },
          "role": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["dastkar-app-creator", "dastkar-admin-user"]
          }
        }
      }
    }
  },
...
}